+++ HeteroSymNN/__init__.py
@@ -1,6 +1,7 @@
+__version__ = "0.3.0b7"
+
 from . import API, Core, JIT, Backend, exceptions
 from .config import settings
 
 
 __all__ = ["API","Core","JIT","Backend","exceptions","settings"]
-__version__ = "0.3.0b7"